Move extension method classes
