+++ Google.Solutions.LogAnalysis.QuickTest/Program.cs
@@ -23,8 +23,8 @@ using Google.Apis.Auth.OAuth2;
 using Google.Apis.Compute.v1;
 using Google.Apis.Logging.v2;
 using Google.Apis.Services;
-using Google.Solutions.IapDesktop.Extensions.LogAnalysis.Extensions;
 using Google.Solutions.IapDesktop.Extensions.LogAnalysis.History;
+using Google.Solutions.IapDesktop.Extensions.LogAnalysis.Logs;
 using System;
 using System.IO;
 using System.Linq;
